migrated to trigger slice
https://coveord.atlassian.net/browse/KIT-548
