+++ packages/headless/src/features/triggers/triggers-state.ts
@@ -1,0 +1,10 @@
+export interface TriggerState {
+  /**
+   * The URL to redirect the user to after receiving a redirection trigger.
+   */
+  redirectTo: string | null;
+}
+
+export const getTriggerInitialState: () => TriggerState = () => ({
+  redirectTo: null,
+});
